windows/netuio: add Windows NetUIO kernel driver
The Windows netuio kernel driver provides the DPDK userspace application
with direct access to hardware, by mapping the HW registers in userspace
and allowing read/write operations from/to the device
configuration space.

Two IOCTLs are defined by the netuio interface:
  * IOCTL_NETUIO_MAP_HW_INTO_USERMODE
      - used for mapping the device registers into userspace
  * IOCTL_NETUIO_PCI_CONFIG_IO
      - used to read/write from/into the device configuration space

Note:
Requests to map the device BARs into userspace need to be processed
in the thread context of the process that initiated the mapping request.
Otherwise, the BARs might end up mapped into an arbitrary process
address space.
EvtIoInCallerContext is used to ensure the requests are handled
in the right user thread context. Other requests (PCI config)
are sent back to the framework and processed by the EvtIoDeviceControl
callback.

Cc: Harini Ramakrishnan <Harini.Ramakrishnan@microsoft.com>
Cc: Omar Cardona <ocardona@microsoft.com>
Cc: Dmitry Malloy <dmitrym@microsoft.com>
Signed-off-by: Narcisa Vasile <navasile@microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>