app/testpmd: fix valid desc id check
The number of desc is a per queue configuration. But in the check function,
nb_txd & nb_rxd are used to check whether the desc_id is valid. nb_txd &
nb_rxd are the global configuration of number of desc. If the queue
configuration is changed by cmdline liks: "port config xx txq xx ring_size
xxx", the real value will be changed.

This patch use the real value to check whether the desc_id is valid. And if
these are not configured by user. It will use the default value to check
it, since the rte_eth_rx_queue_setup & rte_eth_tx_queue_setup will use a
default value to confiure the queue if nb_rx_desc or nb_tx_desc is zero.

Fixes: af75078fece3 ("first public release")
Cc: stable@dpdk.org

Signed-off-by: Chengchang Tang <tangchengchang@huawei.com>
Signed-off-by: Wei Hu (Xavier) <xavier.huwei@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>