vhost/crypto: fix initialization.
This patch fixes the problem that vhost crypto cannot be
initialized due to the different requirement between
built-in virtio-net and virtio-crypto. The fix includes
the following change:

- Added new internal enum type virtio_backend_type to
distinguish virtio-net, virtio-crypto, and external
device types.
- Added new API rte_vhost_crypto_driver_start to
distinguish between virtio-net and virtio-crypto built-in
drivers initialization.
- Added new internal function for the vhost library
to use different feature flags when initializing
virtio-crypto.

Fixes: 2ab58f20db51 ("vhost: refactor virtio ready check")
Cc: maxime.coquelin@redhat.com

Signed-off-by: Fan Zhang <roy.fan.zhang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>