net/bnxt: support fast mbuf free
Add support for DEV_TX_OFFLOAD_MBUF_FAST_FREE to bnxt
vector mode transmit. This offload may be enabled
only when multi-segment transmit is not needed, all
transmitted mbufs for a given queue will be allocated
from the same pool, and all transmitted mbufs will
have a reference count of 1.

Since this offload is only supported in vector mode at
this time, support in the features list is indicated
as "partial".

Signed-off-by: Lance Richardson <lance.richardson@broadcom.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>