net/failsafe: check correct error code while handling sub-device add
When adding a sub-device, it's possible that the sub-device is configured
successfully but later fails to start. This error should not be masked.
The driver needs to check the error status to prevent endless loop of
trying to start the sub-device.

fixes (ae80146c5a1b net/failsafe: fix removed device handling)

cc: stable@dpdk.org
Signed-off-by: Long Li <longli@microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>