net/mlx5: optimize shared counter memory
Currently, when a counter is allocated, the counter list entry memory
will not be used until the counter is released and added back to the
counter container free list. In this case, if a counter is allocated
as shared counter, the shared information can be save to the counter
list entry memory. This adjustment engages to save more memory for
the shared counter.

One more thing is that now the shared counter is only available for
single counter as shared counter may be applied to both root and
none root table, but batch counter with offset is not supported in
root table in some old OFED version. The batch counter with offset
is now fully supported in the current OFED. This commit is also
the initialize change for the batch counter with offeset can be
shared in the later change.

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>